# cmd/api/docs/swagger.yaml
definitions:
...
  types.TestCaseAssert:
    properties:
      flags:
        additionalProperties:
          type: integer
        type: object
      halted:
        type: boolean
      memory:
        additionalProperties:
          type: integer
        type: object
      register:
        additionalProperties:
          type: integer
        type: object
    required:
    - flags
    - halted
    - memory
    - register
    type: object
  types.TestCaseInit:
    properties:
      flags:
        additionalProperties:
          type: integer
        type: object
      memory:
        additionalProperties:
          type: integer
        type: object
      register:
        additionalProperties:
          type: integer
        type: object
    required:
    - flags
    - memory
    - register
    type: object
...
  types.TestCaseRequest:
    properties:
      assert:
        $ref: '#/definitions/types.TestCaseAssert'
      init:
        $ref: '#/definitions/types.TestCaseInit'
      name:
        type: string
    required:
    - assert
    - init
    - name
    type: object
  types.TestCaseResponse:
    properties:
      assert:
        $ref: '#/definitions/types.TestCaseAssert'
      id:
        type: integer
      init:
        $ref: '#/definitions/types.TestCaseInit'
      name:
        type: string
      test_suite_id:
        type: integer
    required:
    - name
    type: object
...
info:
  contact: {}
  description: This is the API documentation for the Assembly Visual project.
  termsOfService: http://swagger.io/terms/
  title: Assembly Visual API documentation
  version: "1.0"
paths:
...
  /classroom/{class_id}/assignment/{assignment_id}/test-suite/{test_suite_id}/test-case:
    get:
      consumes:
      - application/json
      description: Retrieve all test cases by test suite ID
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.TestCaseResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-cases
    post:
      consumes:
      - application/json
      description: Add a new test case
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      - description: Test Case Request
        in: body
        name: test_case
        required: true
        schema:
          $ref: '#/definitions/types.TestCaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-cases
  /classroom/{class_id}/assignment/{assignment_id}/test-suite/{test_suite_id}/test-case/{test_id}:
    delete:
      consumes:
      - application/json
      description: Delete a test case
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      - description: Test Case ID
        in: path
        name: test_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-cases
    get:
      consumes:
      - application/json
      description: Retrieve a test case by its ID
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      - description: Test Case ID
        in: path
        name: test_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TestCaseResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-cases
    put:
      consumes:
      - application/json
      description: Update an existing test case
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      - description: Test Case ID
        in: path
        name: test_id
        required: true
        type: integer
      - description: Test Case Data
        in: body
        name: test_case
        required: true
        schema:
          $ref: '#/definitions/types.TestCaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-cases
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"